+++ core/types/user.go
@@ -17,7 +17,7 @@ type SubsonicUser struct {
 	PodcastRole         bool   `json:"podcastRole" xml:"podcastRole,attr"`                 // Optional: Manage podcasts. Default: false
 	ShareRole           bool   `json:"shareRole" xml:"shareRole,attr"`                     // Optional: Share files. Default: false
 	VideoConversionRole bool   `json:"videoConversionRole" xml:"videoConversionRole,attr"` // Optional: Start video conversions. Default: false
-	MaxBitRate          int64  `json:"maxBitRate" xml:"maxBitRate,attr"`                   // Optional: Maximum bitrate for streaming. Default: 0 (no limit)
+	MaxBitRate          int    `json:"maxBitRate" xml:"maxBitRate,attr"`                   // Optional: Maximum bitrate for streaming. Default: 0 (no limit)
 }
 
 type SubsonicUsers struct {
@@ -25,7 +25,7 @@ type SubsonicUsers struct {
 }
 
 type User struct {
-	Id                  int64  `json:"id"`
+	Id                  int    `json:"id"`
 	Username            string `json:"username" xml:"username"`                       // Required: The name of the new user.
 	Password            string `json:"password" xml:"password"`                       // Required: The password of the new user, either clear text or hex-encoded.
 	Email               string `json:"email" xml:"email"`                             // Required: The email address of the new user.
@@ -43,6 +43,6 @@ type User struct {
 	PodcastRole         bool   `json:"podcastRole" xml:"podcastRole"`                 // Optional: Manage podcasts. Default: false
 	ShareRole           bool   `json:"shareRole" xml:"shareRole"`                     // Optional: Share files. Default: false
 	VideoConversionRole bool   `json:"videoConversionRole" xml:"videoConversionRole"` // Optional: Start video conversions. Default: false
-	MaxBitRate          int64  `json:"maxBitRate" xml:"maxBitRate"`                   // Optional: Maximum bitrate for streaming. Default: 0 (no limit)
+	MaxBitRate          int    `json:"maxBitRate" xml:"maxBitRate"`                   // Optional: Maximum bitrate for streaming. Default: 0 (no limit)
 	Folders             []int  `json:"folder" xml:"folder"`                           // Optional: IDs of music folders the user can access.
 }
